Fix blinking LED during confirmation request
Due to recent changes in Trussed, the LED is blinking during a user
confirmation request if it is processing data, for example on
webauthn.bin.coffee.  This patch restores the old behavior by using a
custom Trussed version that does not change the LED when processing data
and by fixing the refresh method in board/src/trussed.rs.