usb-ccid: Implement abort handling
Previously, we just panicked when receiving an abort request.  This
caused a long delay in scdaemon/gnupg when the NK3 was connected.  With
this patch, we add abort handling to usb-ccid to fix this issue.

The implementation is based on Section 5.3.1 of the CCID specification
[0].  The device is expected to send two abort messages, one over the
bulk endpoint and one over the control pipe.  We perform the abort once
we have received both messages.  In the time between receiving the first
and the second message, we reject all incoming messages on the bulk
endpoint.

Fixes https://github.com/Nitrokey/nitrokey-3-firmware/issues/22

[0] https://www.usb.org/sites/default/files/DWG_Smart-Card_CCID_Rev110.pdf