Generate commands.bd automatically
Previously, we manually updated the commands.bd for the lpc55 firmware
signature process.  With this patch, we automate this process:  We parse
the Cargo manifest to determine the current firmware version and the
build profile to determine the filesystem boundary.  We then generate
the commands.bd file based on this data.