Create workspace for components, embedded runner, utils
This patch creates a virtual workspace for all components, the embedded
runner and the utils.  This makes sure we use the same dependency
versions and overrides in these crates.  The other runners are not
included because of dependency clashes with the embedded runner.

Note that you might have to clean up your components directory if you
have old subdirectories that do not contain a crate.