embedded/lpc55: Disable NFC for hardware checks
This patch disables the NFC functionality if hardware checks are
enabled.  This is because of the issue that we currently cannot talk to
both the external flash and the NFC chip with the same SPI
configuration.  This does not affect end users as the hardware checks
are only executed during provisioning.