embedded/lpc55: Use external flash in USB mode
This patch changes the external storage for the lpc55 SoC in the
embedded runner to the external flash unless the device is in passive
NFC mode.  In this case, 256 bytes of RAM are used to simulate the
external flash.  This makes sure that the littlefs header can be written
(though formatting fails with a NoSpace error).

As we still need to initialize the NFC chip, we now do that in the
provisioner (and no longer check the external flash).  Instead the
external flash is checked on the first boot into the regular firmware.