ext-flash: layout properties; debug consistency
The external flash now has a reserved portion at the end of
the flash. 128kb (32 blocks) are left out at the end. This
is done by using a reduced BLOCK_COUNT, which will restrict
littlefs2 to the given BLOCK_COUNT, but the low-level
read/write/erase methods for `ExtFlashStorage` will still
accept operations on this region.