lpc55-builder: Add provision-release target
This patch adds the provision-release target to the lpc55-builder
Makefile that provisions the keystore and enables the encrypted storage
using the PRINCE peripheral.  This setup is closer to production than
our develop setup without storage encryption.  The advantage of the
develop setup is that it is easy to read out the internal flash for
debugging.