ci: Separate normal and release jobs
This patch separates the jobs executed for normal builds (push, nightly)
and for release builds (web):  For normal builds, only the source code
is checked but no binaries are built.  For release builds, only the
required binaries are built.  This should speed up the builds and reduce
confusion between stable and alpha images.

As part of these changes, this patch also:
- adds a check that releases are always created from a proper tag,
- drops the zip files for the binaries,
- changes the naming so that the binaries are always called firmware and
  provisioner (not: alpha), and
- enables the sparse protocol for the crates.io registry.