usbip: Add interactive user presence check
This patch adds a setting for the usbip runner, user presence.  The
default and current behavior is to always accept user presence checks.
Alternatively, they can always be rejected or answered interactively
with a prompt.  To make this work, we need to update the trussed-usbip
crate to support keepalive messages.