lpc55: Only enable PRINCE for the FS region
Previously, we enabled the complete PRINCE region 2 if we wanted to
access the FS.  This means that we also tried to encrypt parts of the
firmware if was written after 0x80_000 (start of region 2) although the
firmware is not encrypted.

With this patch, we only enable the PRINCE subregions that actually
contain the filesystem.  This requires extracting the filesystem
definition from the lpc55-hal crate.