lpc55: Move USB initialization to end
This patch splits the NFC and USB initialization and moves the USB
initialization to the end of the initialization process on the lpc55.
This makes sure that we can respond to all USB requests as the poll task
is started directly after the init process.

Previously, we initialized the USB interface at the beginning of the
process so that we could not respond to requests until the end of the
init process.  This could potentially mean that the device would lose
the initial GET DESCRIPTOR request sent by the Linux kernel, causing a
delay of typically 5 s after connecting the device.

Fixes: https://github.com/Nitrokey/nitrokey-3-firmware/issues/302