Refactor versioning
This patch changes the way we handle versioning:  The crate version
defined in the root Cargo.toml is now only changed for stable releases
or release candidates.  The test suffix is automatically added if the
test feature is enabled.  Additionally, build metadata with the Git
commit and worktree status is added to the version string.  Finally,
there are additional checks to make sure that tag names match the
computed version.

Fixes: https://github.com/Nitrokey/nitrokey-3-firmware/issues/339