Add nix build tooling
This patch adds a nix-shell environment and a Makefile to build the
firmware binaries using Nix.  The idea is to make it easier to build a
previous version of the firmware the same way it was built at the time
of development.  Potentially, it could also be used with the CI for
reproducible builds (to be investigated).