Handle config error in admin app
This patch updates the admin-app dependency to report errors when
loading the config from IFS during intialization.  It adds a new init
status flag CONFIG_ERROR and disables the opcard application if such an
error occured to ensure security and stability (as the opcard
configuration will determine the backend that is used).