Remove unused boards and profiles
This patch removes boards that are no longer supported, namely nrfdk and
proto1 using the NRF52.  It also removes the nrf52 build profile that is
no longer needed.  And boards that have already been dropped are now
also removed from the lpc55 profile.

Finally, this patch also removes the qspi implementation for NRF52 and
the extflash feature flags because we always use the spi implementation.