Move memory regions into component
The memory regions are not only required by the runner, but also by the
gen-commands-bd util.  Therefore this patch moves them into a component
that can be referenced directly by both crates.  This also allows us to
drop the code generation in the runner build script and directly use the
constants from the memory regions component in the firmware.