nk3am: Remove blocking UP check
Since this commit [0] that changed the threshold from 3 to 1, the
blocking loop in the user presence check for the NK3AM does no longer
have any functional effect.  But it has undesired side effects, for
example skipping the UI update for a second so that the blinking effect
is not triggered, and delaying the handling of cancelled requests.

This patch removes the unnecessary blocking loop.  The button
implementation still requires 100 ticks to trigger a button press.

Fixes: https://github.com/Nitrokey/nitrokey-3-firmware/issues/93

[0] https://github.com/Nitrokey/nitrokey-3-firmware/commit/35dac8d850c007cb0bb495c51085b9f2b6a39b06