Update ctap-types and fido-authenticator
This patch updates ctap-types so that it uses the COSE implementation in
the cosey crate instead of shipping its own identical cose module.  This
reduces code duplication and has a small effect on the binary size.  We
also need to update fido-authenticator to use cosey instead of
ctap_types::cosey, and we need to update cose so that it includes a fix
that we previously applied to the cose module in ctap-types.

Additionally, many other optimizations of ctap-types and
fido-authenticator are pulled in with this patch.