lpc55-builder: Use GDB server CLI
Previously, we used the JLinkGDBServer binary that links against the
outdated Qt4.  This patch replaces it with JLinkGDBServerCLExe that is
included in the same bundl to get rid of the EOL Qt4 dependency and to
make sure that we always use CLI mode. (JLinkGDBServer sometimes runs in
GUI mode and sometimes in CLI mode depending on the exact version and
setup.)