Move some statics into init task
To avoid unsafe code and fix some clippy lints, we can move static muts
into the locals for the init task which also have static lifetime.  This
is currently not possible for all statics because some need to be
accessed globally.  This patch moves the easy cases into a new Resources
struct and adds it to the locals of the init task.