Allow missing mechanisms in usbip runner
Some mechanisms are only implemented by the SE050 backend and not
available in the usbip runner.  Instead of failing the compilation,
this patch ignores those mechanisms when compiling the usbip runner so
that a runtime error is raised when such a mechanism is used.  In the
medium term, we should use a backend providing a software implementation
of these mechanisms.