ci: Move metrics comment from after_script to script
Previously, we had the metrics comparison and comment creation in the
after_script part of the metrics job so that a failure in these commands
would not cause a CI failure.  As we increasingly rely on the metrics
comments and the relevant commands are pretty stable, this patch moves
them to the script part so that we will notice failures more easily.