ci: Only comment metrics on other branches than main
In https://github.com/Nitrokey/nitrokey-3-firmware/pull/594, we changed
the metrics job to fail if the comment cannot be posted.  This causes a
CI failure on main as there is no matching PR for the CI run.  This
patch adds a condition to only post the comment if the CI runs against
a branch other than main.