ci: Only generate metrics comment for other branches than main
The old logic can lead to problems when the job runs for the first time
for a commit on main, e. g. after a merge with a merge commit (instead
of a fast-forward merge), as the "root" of the branch is equal to the
head and there are no cached metrics for the head.

This also reverts the temporary change to ignore errors during comment
creation in https://github.com/Nitrokey/nitrokey-3-firmware/pull/600 as
the cause should be fixed with this PR.