provisioner: Remove ReformatFilesystem command
The ReformatFilesystem command in the provisioner app required direct
access to the IFS storage.  This is potentially unsound and adds
complexity to the store implementation.  To avoid this requirement, this
patch removes the command.  Instead a factory reset can be performed
via admin-app to clear the filesystems.  If they need to be formatted,
a custom firmware should be used that can format the filesystems during
the store setup.