Require Nitrokey SDK v0.3.0 in manifest
This patch replaces the pynitrokey version requirement with a Nitrokey
SDK version requirement.  See this PR for more information on the
mechanism:
- https://github.com/Nitrokey/nitrokey-sdk-py/pull/65

We also bump the pynitrokey version requirement to make sure that old
versions that are not aware of the SDK version requirement reject new
updates.

The reason for this new requirement is the support for the migration to
the filesystem layout v2, see:
- https://github.com/Nitrokey/nitrokey-sdk-py/pull/67
- https://github.com/Nitrokey/nitrokey-3-firmware/pull/602